requirements.txt
django==1.7.3
django-cms==3.0
cmsplugin-filer